ECM-Projects/transfer-master/ideas_project/requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==2.2.8
django-crispy-forms==1.8.1
django-markdown-deux==1.0.5
django-microsoft-auth==2.3.1
django-model-utils==4.0.0
django-notifications-hq==1.5.0
django-pagedown==2.0.3
djangorestframework==3.10.3
idna==2.8
jsonfield==2.0.2
markdown2==2.3.8
oauthlib==3.1.0
Pillow==6.2.1
pycparser==2.19
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
